Added 32-bit fixed-point numeric types.

+++ Foundations/Extensions/MathStatic.cs
@@ -14,6 +14,7 @@ FILE GENERATOR. IF YOU SAVE THE FILE IN VISUAL STUDIO IT WILL DO THIS FOR YOU.
 */
 
 using System;
+using Foundations.Types;
 
 namespace Foundations
 {
@@ -70,45 +71,45 @@ namespace Foundations
     }
 
     /// <summary>
-    /// Implementations of <see cref="System.Math"/> methods for <see cref="System.Single"/> arguments.
+    /// Implementations of <see cref="System.Math"/> methods for <see cref="Single"/> arguments.
     /// </summary>
     public static partial class MathF
     {
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Acos.</summary>
-        public static Single Acos(Single x) => (Single)Math.Acos(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Acos.</summary>
+        public static Single Acos(Single x) => (Single)Math.Acos((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Asin.</summary>
-        public static Single Asin(Single x) => (Single)Math.Asin(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Asin.</summary>
+        public static Single Asin(Single x) => (Single)Math.Asin((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Atan.</summary>
-        public static Single Atan(Single x) => (Single)Math.Atan(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Atan.</summary>
+        public static Single Atan(Single x) => (Single)Math.Atan((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Cos.</summary>
-        public static Single Cos(Single x) => (Single)Math.Cos(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Cos.</summary>
+        public static Single Cos(Single x) => (Single)Math.Cos((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Cosh.</summary>
-        public static Single Cosh(Single x) => (Single)Math.Cosh(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Cosh.</summary>
+        public static Single Cosh(Single x) => (Single)Math.Cosh((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Exp.</summary>
-        public static Single Exp(Single x) => (Single)Math.Exp(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Exp.</summary>
+        public static Single Exp(Single x) => (Single)Math.Exp((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Log10.</summary>
-        public static Single Log10(Single x) => (Single)Math.Log10(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Log10.</summary>
+        public static Single Log10(Single x) => (Single)Math.Log10((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Sin.</summary>
-        public static Single Sin(Single x) => (Single)Math.Sin(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Sin.</summary>
+        public static Single Sin(Single x) => (Single)Math.Sin((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Sinh.</summary>
-        public static Single Sinh(Single x) => (Single)Math.Sinh(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Sinh.</summary>
+        public static Single Sinh(Single x) => (Single)Math.Sinh((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Sqrt.</summary>
-        public static Single Sqrt(Single x) => (Single)Math.Sqrt(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Sqrt.</summary>
+        public static Single Sqrt(Single x) => (Single)Math.Sqrt((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Tan.</summary>
-        public static Single Tan(Single x) => (Single)Math.Tan(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Tan.</summary>
+        public static Single Tan(Single x) => (Single)Math.Tan((double)x);
 
-        /// <summary><see cref="System.Single"/> wrapper for System.Math.Tanh.</summary>
-        public static Single Tanh(Single x) => (Single)Math.Tanh(x);
+        /// <summary><see cref="Single"/> wrapper for System.Math.Tanh.</summary>
+        public static Single Tanh(Single x) => (Single)Math.Tanh((double)x);
 
     }
 
@@ -117,40 +118,40 @@ namespace Foundations
     /// </summary>
     public static partial class MathExtensions
     {
-        /// <summary></summary>
+        /// <summary>Acos</summary>
         public static Single Acos(this Single x) => MathF.Acos(x);
 
-        /// <summary></summary>
+        /// <summary>Asin</summary>
         public static Single Asin(this Single x) => MathF.Asin(x);
 
-        /// <summary></summary>
+        /// <summary>Atan</summary>
         public static Single Atan(this Single x) => MathF.Atan(x);
 
-        /// <summary></summary>
+        /// <summary>Cos</summary>
         public static Single Cos(this Single x) => MathF.Cos(x);
 
-        /// <summary></summary>
+        /// <summary>Cosh</summary>
         public static Single Cosh(this Single x) => MathF.Cosh(x);
 
-        /// <summary></summary>
+        /// <summary>Exp</summary>
         public static Single Exp(this Single x) => MathF.Exp(x);
 
-        /// <summary></summary>
+        /// <summary>Log10</summary>
         public static Single Log10(this Single x) => MathF.Log10(x);
 
-        /// <summary></summary>
+        /// <summary>Sin</summary>
         public static Single Sin(this Single x) => MathF.Sin(x);
 
-        /// <summary></summary>
+        /// <summary>Sinh</summary>
         public static Single Sinh(this Single x) => MathF.Sinh(x);
 
-        /// <summary></summary>
+        /// <summary>Sqrt</summary>
         public static Single Sqrt(this Single x) => MathF.Sqrt(x);
 
-        /// <summary></summary>
+        /// <summary>Tan</summary>
         public static Single Tan(this Single x) => MathF.Tan(x);
 
-        /// <summary></summary>
+        /// <summary>Tanh</summary>
         public static Single Tanh(this Single x) => MathF.Tanh(x);
 
     }
